version=12.1.7_dev # new feature: Vasp reader shows enthalpy and Gibbs energy in model name

+++ Jmol/src/org/jmol/adapter/readers/xml/XmlVaspReader.java
@@ -126,9 +126,9 @@ public class XmlVaspReader extends XmlReader {
       atomSetCollection.setDoFixPeriodic();
       atomSetCollection.newAtomSet();
       if (enthalpy != null)
-        atomSetCollection.setAtomSetAuxiliaryInfo("enthalpy", Float.valueOf(enthalpy));
+        atomSetCollection.setAtomSetAuxiliaryInfo("enthalpy", Double.valueOf(enthalpy));
       if (gibbsEnergy != null)
-        atomSetCollection.setAtomSetAuxiliaryInfo("gibbsEnergy", Float.valueOf(gibbsEnergy));
+        atomSetCollection.setAtomSetAuxiliaryInfo("gibbsEnergy", Double.valueOf(gibbsEnergy));
       if (enthalpy != null && gibbsEnergy != null)
         atomSetCollection.setAtomSetName(" Enthalpy = " + enthalpy + " Gibbs Energy = " + gibbsEnergy);
       return;
